Ensure resolved final methods are called directly
In 3b2830ef genInvokeVirtual() was modified to detect resolved private
methods for nestmates. If a private method is found, it generates a
direct call, and otherwise (including when the method reference is
unresolved) it generates a virtual call. In order to generate a virtual
call, it queries the constant pool a second time, and if the constant
pool entry was originally unresolved, it might have become resolved
since. In this case, ilgen could fail to identify for a resolved call
that the called method is final, and generate an indirect instead of a
direct call. Normally it's harmless for the call to be indirect, but
there are a number of final methods recognized by the JIT compiler that
undergo transformations assuming that calls to them are direct.

To prevent this problem, genInvokeVirtual() will now re-read method from
the result of findOrCreateVirtualMethodSymbol() whenever that result is
resolved.

Signed-off-by: Devin Papineau <devinmp@ca.ibm.com>