Fix crash in AOT during field iteration
The code in the constructor for TR_VMFieldsInfo was making the
assumption that the superclass for a class can always be found. While
this is always true when JIT compiling a method, it is possible under
certain rare circumstances for this to fail during an AOT compilation.
For instance, this can fail under SVM if the SCC runs out of space
while trying to store the class chain for the parent class.

When an unexpected failure was encountered, this would then cause the
JVM to crash. This has now been fixed so that under AOT this will cause
the compilation to fail instead of crashing the JVM.

Fixes: #4751
Signed-off-by: Ben Thomas <ben@benthomas.ca>