Do not ignore any relocation records
Currently there is a coupling between the existence of guards and the
relocation of the inlined table in the metadata. This means that if
the JIT does not produce a guard for certain inlined sites, the
associated entry in the inlined table in the metadata will not be
relocated. This has the consequence of causing other relocations that
need information via this entry to not have the means to do so. The
only way to proceed is to fail the AOT load - if the compiler cannot
materialize a value to relocate a location that **MUST** be
relocated, the code **CANNOT** be loaded and executed.

Signed-off-by: Irwin D'Souza <dsouzai@ca.ibm.com>