Fix issues with thunks in remote AOT mode
[skip ci]
JITaaS generates all thunks inside the code cache and later
relocates them in one batch on the client at the end of the compilation.
This means get/set/persist thunk logic should be slightly different
from mainline.
We do not need shared cache versions of methods like `setJ2IThunk`
on the server, because client already receives all thunks,
so we can store them to shared class cache at the time of relocation.
Similarly, we do not need shared cache version of `getJ2IThunk`, because
during compilation all thunks are on the server, and haven't been
written to shared cache yet. They can be looked up
in the same way as for non-AOT.

To this end,
- Remove `TR_J9SharedCacheServerVM` versions of methods that
get/set/persist thunks.
- In `TR::CompilationInfoPerThread::relocateThunks()` call the shared
cache version of `setJ2IThunk` in addition to the base version.
- Override `TR_ResolvedRelocatableJ9JITaaSServerMethod::constantPool()`
to match `TR_ResolvedRelocatableJ9Method` implementation. It needs to
return ROM literals and without it thunk relocation will not work.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>