Add AOT upgrades based on SCC hints to the LPQ
Forced AOT upgrades are recompilations that are based on hints stored in
the SCC. These hints are used to identify methods that were AOT compiled
(or loaded), but were recompiled because the sampling mechanism
determined that they were important methods. However, these upgrades are
added immediately to the queue when the method is loaded from the SCC,
which can potentially impact an application's rampup phase.

This commit defers these upgrades until free cycles are available.

Signed-off-by: Irwin D'Souza <dsouzai@ca.ibm.com>