Marshal platform specific relo data
The data that is stored in the _targetAddress/_targetAddress2 may be
different on different platforms. However, relocation header data that
is stored to to the SCC is common across platforms. In order to fully
consolidate the code that writes the header information, this commit
introduces a method that has to be implemented on each platform that is
used to demarshal the platform specific data from the
TR::IteratedExternalRelocation object and marshal it into the platform
agnostic form that is used to write out the header information.

Signed-off-by: Irwin D'Souza <dsouzai@ca.ibm.com>