Enable SVM for remote AOT
[skip ci]
Make changes to Symbol Validation Manager that allow
to use it for remote AOT compilations.
- Replace J9_CLASS_FROM_CP macros with front-end calls.
- Initialize server vm for SVM
- SVM has a map of symbols to their respective ids - `_symbolToIdMap`.
This map is populated during compilation on the server, but is also
used during validation on the client. Thus, at the end of the
compilation we need to send it from server to the client.
- Some methods of `TR_ResolvedJ9Method` and `TR_J9VM` (non-AOT versions)
create SVM records when these methods are called during AOT compilation.
So we need to do the same for base server implementation of these
methods.
- Create new queries for `TR_J9SharedCache` and `TR_J9VM` which are
needed for remote calls

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>