Add support to compress AOT code
For any AOT compiled method we store AOT compiled code with
metadata to aid relocation in AOT load run.
This commits adds sequence to compress the data before storing it into
cache to reduce the size of SharedClass Cache using zlib compression
library. While loading method from the cache, it decompressed the data
for method before it applied relocation.

This commit also enables this feature by default on z platform.
On Linux on Z it used zlib library that comes with OpenJ9  to deflate/inflate
the data. On zOS it links zlib installed on build machine system with JIT.

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>