Make String.indexOf(String) use intrinsics whenever possible
This commit fixes a bug in String.indexOf(String) API where certain optimal
paths (intrinsics that can be hardware accelerated) can never be reached.
As a result of this, the String.indexOf(string) API always defaults to the sub-optimal
string search logic.
This applies to both compact and non-compact string modes.

Signed-off-by: Nigel Yu <yunigel@ca.ibm.com>