Use client's ROM class pointer in a front-end query
[skip ci]
`TR_J9VM::getPersistentClassPointerFromClassPointer` returns a pointer
to a ROM class. However, on the server, we store a copy of client's
ROM class, and calling this method would return a pointer to that copy.
But this method is always used to measure the offset of a ROM class
from the beginning of shared cache.
Thus, we need to get a pointer to the ROM class on the client side.
Do it by adding a pointer to ClassInfo struct, so that for all cached
ROM classes we know its pointer on the client.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>