Fixed freeing of memory for cudaInfo for JIT GPU
In the single GPU kernel case, regionExitGPU calls freeGPUScope
which frees the memory for the cudaInfo object. The code then
calls getStateGPU with the exact same cudaInfo object that was
just freed. This is now changed so that the cudaInfo object is freed
inside getStateGPU instead.

regionExitGPU would also flush data to the CPU in some cases even after an
error case has been triggered. This was changed to no longer happen.

This change also cleans up the code by removing an unused parameter to
regionExitGPU and changing int return types from NULL_DEVICE_PTR to 0.

Issue: #5013
Signed-off-by: jimmyk <jimmyk@ca.ibm.com>