Enable Concurrent Scavenger on BE Linux for Power
This change enables the OMR_GC_CONCURRENT_SCAVENGER flag for Big Endian
Linux build for Power. This flag enables the use of concurrent scavenger GC
via the -Xgc:concurrentScavenge option.

This also updates the _sp builds. They are supposed to have the same flags
as the normal spec so some of them now need the
OMR_GC_CONCURRENT_SCAVENGER flag to be enabled.

Issue: #5023
Signed-off-by: jimmyk <jimmyk@ca.ibm.com>