Update options for compilation request
A new option object is created for each compilation in `wrappedCompile()`.
The new option object is constructed from `_aotCmdLineOptions`
instead of `_jitCmdLineOptions` when AOT is used. However options for jitaa
were only set in `fePostProcessJIT()` but not in `fePostProcessAOT()`.
Add a helper method `setupJITaasOptions()` to be called by both JIT and AOT
option post processes.

Fixes: #5016

Signed-off-by: Annabelle Huo <Annabelle.Huo@ibm.com>