Always use non-AOT front-end in packRemoteROMClassInfo
[skip ci]
Previously, if `packRemoteROMClassInfo` was called
during AOT compilation, all of the values for class info
would be obtained by calling front-end methods of a
Shared Cache VM.
This is a problem, because shared cache versions of the
methods attempt to validate classes and may return NULL.
This is wrong for 2 reasons:
1. We perform validation on the server.
2. Values packed in `packRemoteROMClassInfo` are cached
on the server, for the entire lifetime of a client.
If a shared cache version of a method sets class info
parameter to NULL, then it will remain NULL on the server,
even for non-AOT compilations. This is wrong, because
non-AOT requires most parameters to be non-NULL.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>