Fix whitespace and return values in GPUHelpers
Many of the functions inside GPUHelpers.cpp returned magic numbers or
results that did not match the return type for no reason. This fixes the
functions that return magic numbers to return a macro that says what the
magic number means. It also fixes functions to return the proper type.

isGPUFailed was renamed checkForGPUFailure to be more descriptive.

This also fixes the whitespace in many places in the file.

Issue: #5215
Signed-off-by: jimmyk <jimmyk@ca.ibm.com>