Handle delayed AOT load for JITaaS Remote AOT
AOT compilations can fail on purpose because we want to load the AOT body later on. This case is signalled by having a successful compilation but canRelocateMethod == false. When we get a delayed AOT load, we want to manipulate the invocation count (just like how non-JITaaS AOT handles it). Also, logCompilationSuccess requires additional handling as this special case has a NULL metaData.
[skip ci]
Issue: #5131

Signed-off-by: Harry Yu <harryyu1994@gmail.com>