Do not cache array class of RAM class
[skip ci]
Array class of RAM class can be initialized
when the class is already loaded, so it is not
safe to cache NULL as array class.
Only cache non-NULL values, make remote call otherwise.
If a remote call returns non-NULL, it means that array
class has been initiailized. It should not change anymore
so it's safe to cache it.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>