Dynamic checkcast evaluation on x86
Currently on X checkcast evaluator only supports inline evaluation for
constant cast classes. If the cast class is not known at compile time a VM
helper call is generated. With upcoming optimizer changes where certain
instanceOf nodes are transformed into checkcast with dynamic cast class,
there is a need to support the inline evaluation for these new cases. This
PR provides such support. It inlines the checkcast instructions needed for
interface and class case; for array case, a VM helper call is generated,
but this is the same as in constant cast class case since array cases are
believed to be uncommon. As the cast class is dynamic, runtime tests need
to be generated to tell if the cast class is array, interface, or class.
For interface, a tight loop through the itable is generated; for class,
the necessary equality test and superclass test are generated.
Consideration was made to add some form of cache slot similar to the
constant case; but given that the cast class is dynamic, 1) instead of
just 1 slot, we would need 2 slots to match both the cast class and the
cached value in order to produce a hit, and 2) given the dynamic nature
where the cast class will potentially vary every time the checkcaast is
executed, the hit rate for the cached pair are likely low. So this as an
initial support only inlines the necessary tests without the cached slots,
which can be added as a followup change in the future.

Signed-off-by: Yan Luo <Yan_Luo@ca.ibm.com>