Handle mismatched env setup between server and client
[skip ci]
When `tryCompilingAgain`, we do JIT compilation instead of using AOT.
If `Xnojit` is set, retry should not go through. Meanwhile with JITaaS,
the client should not send a remote compilation request.
Otherwise, the server would do a JIT compilation which
contradicts `-Xnojit` set at the client.
This change only includes the client side change. The next step
is to remove `TR_EnableJITaaSRemoteAOT` env flag so that the server
is alway eligible for AOT compilation when the client requests.

Issue #5163

Signed-off-by: Annabelle Huo <Annabelle.Huo@ibm.com>