Fix vectorized indexOf for negative bytes on Power
Due to the way that sign- and zero-extension of byte and char values
works on Power, the target character passed into the intrinsicIndexOf
helpers was sometimes sign-extended and sometimes not. However, the
evaluators responsible for generating the vectorized code for these
helpers was assuming that these values would always be zero-extended.
This was specifically causing problems when searching for bytes with
negative values, as they were sometimes sign-extended.

This has been fixed by making the evaluators always mask off the upper
bits before performing the initial scalar comparison, which is the only
part of the code which is making this assumption.

Signed-off-by: Ben Thomas <ben@benthomas.ca>