Fix segfaults in vectorized indexOf on Power
Previously, the code for inlining intrinsicIndexOf on Power was making
the assumption that the upper 32 bits of the registers containing the
array indices were well-defined (i.e. either zero extended or sign
extended). As it turns out, many 32-bit operations on Power can cause
the upper 32 bits of a register to be undefined. By making use of those
bits, we were causing undefined behaviour and, in some cases, would read
far beyond the end of the array, resulting in a segfault.

This has been corrected and the evaluator responsible for inlining the
vectorized versions of these methods will now correctly sign extend the
indices before making use of them.

Signed-off-by: Ben Thomas <ben@benthomas.ca>