Re-enable vectorized String.indexOf on 64-bit Power
Previously, the vectorized helpers for String.indexOf were disabled on
Power due to some difficult to reproduce segfaults that they appeared to
be causing. Now that these have (hopefully) been resolved, it should now
be safe to re-enable the vectorized versions of the helpers.

Note that these have only been re-enabled on 64-bit. These versions were
never intended to run on 32-bit and should not have been enabled.

Signed-off-by: Ben Thomas <ben@benthomas.ca>