Return NULL instead of throwing when the class name is invalid
Prior to #4919, ClassLoader.defineClassImpl() would check if the given
class name and immediately throw NoClassDefFoundError if it was invalid.
Beginning with Java 12, JavaLangAccess.defineClass() is used to define
accessor classes (with invalid names). The name validation must still
be done, but throwing the exception must be delayed until it can be
determined to be the appropriate response. The invalid name is signalled
with a bit in the options parameter to defineClassCommon() which must
not set a pending exception in that case.

Signed-off-by: Keith W. Campbell <keithc@ca.ibm.com>