Move implementation of generateBinaryEncodingPrologue on Z to OMR
The `generateBinaryEncodingPrologue` has generic code for the following:

1. Creating the recompilation preprologue
2. Creating the method preprologue
3. Creating the method prologue

On Z there is nothing specific to OpenJ9 about the above steps. In fact
the steps are missing from OMR and hence no prologue/preprologue is
generated of system linkages at the OMR level during binary encoding.
This is a problem and the fix is to simply migrate this code over.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>