Fix performance with delayed relocations
[skip ci]
Previously, when client wanted to delay relocations
after remote AOT compilation, its meta data would be NULL.
This is incorrect, because control is able to tell if
relocations need to be delayed by checking that relocations
failed but meta data stil exists.
If meta data is NULL, the compilation is treated as a failure,
thus AOTed code never gets loaded from SCC during the same run.
The solution is to temporarily create meta data on the client
for this case.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>