Cache field and static attributes per RAM class
[skip ci]
- We used to cache field and static attributes per resolved method,
but these methods return attributes of a RAM class field at a
given CP index. This means that we can cache them per RAM class,
in `ClassInfo` for all cases except when field is unresolved.
If it's unresolved it may become resolved later on, so in this
case we still cache to a per-resolved method cache.
This significantly reduces the number of remote messages.
- `TR_ResolvedRelocatableJ9JITaaSServerMethod::field(static)Attributes`
used the same caches as non-AOT versions of the method. This didn't
cause problems before, because caches existed only for 1 compilation.
However, now that caches exist for multiple compilations,
this is wrong because an AOT compilation might be followed by a non-AOT
compilation that requests attributes for the same field, and we'll get
a wrong value.
Created separate caches for AOT attributes.
- Changed <include> hierarchy so that `JITaaSCompilationThread.hpp`
includes `j9methodServer.hpp`.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>