Keep iTable indices stable in fast HCR
When interface methods were reordered during fast HCR, the VM was not
maintaining the stability of the indices into the iTable. Some uses of
the indices were fixed by the HCR, but several places were missed.

Rather than attempt the missed fixups, keep the iTable indices stable so
nothing needs to be fixed. This is done by maintaining a mapping table
in replaced interface classes which describes the original method order
(in fast HCR methods can not be added or removed, only reordered).

Update the interface method reordering test (rc011) to better test the
JIT behaviour. The test now fails on unpatched VMs.

[ci skip]

Signed-off-by: Graham Chapman <graham_chapman@ca.ibm.com>