Allow concurrent hot compilations in JITaaS mode
The heuristics in the JIT don't allow multiple expensive compilation to be performed in parallel to avoid increasing the peak footprint too much. The default value is set such that hot, scorching and JSR292 compilations are considered expensive. However, in JITaaS world such compilation are performed remotely and therefore there is no risk of blowing the footprint limit on the JVM. By setting a large value for "expensiveCompWeight", we are effectively allowing several memory expensive compilations to be performed concurrently.

Issue: #4476
[skip ci]

Signed-off-by: Harry Yu <harryyu1994@gmail.com>