Rework field/staticAttributes for JITaaS
[skip ci]
- Reimplement `TR_ResolvedRelocatableJ9JITaaSServerMethod::field/staticAttributes`
to use the same set of attributes sent from client to the server as
non-AOT version.
This gets rid of a union and makes reusing code for AOT and non-AOT
versions easier.
- Reimplement `TR_J9MethodFieldAttributes`.
First, it's not a union of 2 structs anymore, because of the previous
point.
Second, instead of containing a union for with either `uint32_t` or
`void *`, it is now just `void *`, since pointer size is big enough
to contain a 4-byte unsigned integer. This makes implementing
`operator==` easier.
Third, implement helper methods for setting attributes and override
`operator==`. The operator override will be useful later, when we'll
be comparing cached attributes with newly fetched attributes.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>