Version check between client and server
Client and server needs to be compatible.
Any kind of uncompatiblity will lead to the crash.
To ensure this, there will be a version check
between client and server. First message after every
connection will have client version which will be
checked on server. On passing the version check server
will go on and read the remaining message.

On failing the version check, connection on the
server will be closed and error code will be send
to client and client will do local compilations.

Signed-off-by: Satbir Singh <satbir.singh1@ibm.com>