Async handler for Concurrent Scavenger
Introduced an Async handler that will be later used by CS to force
mutator thread to promptly flush its copy cache to the scan queue. The
handler is to be activated toward the end of CS cycle when scan queue is
empty (yet to be implemented).

Also, GC registered for two VM hooks, that will help mutator threads
flush their copy caches by themselves

The async handler and hooks are empty at the moment, but will eventually
call a flush-copy-cash method this is yet to be introduced.

The only functional change is: moved VMAccess release hook trigger in
JNICriticalRegion::releaseAccess() to an earlier point to be consistent
with internalReleaseVMAccessNoMutexNoCheck

Minor cleanup: cached J9VMThread in EnvironmentDelegate, since it's
re-fetched about a dozen of times at various spots.

Some background: https://github.com/eclipse/omr/issues/4787

Signed-off-by: Aleksandar Micic <amicic@ca.ibm.com>