Alloc deviation boost option for CS tilting
Just an option to control concurrentScavengerAllocDeviationBoost
parameter from https://github.com/eclipse/omr/pull/3959.

Default value of the parameter is set to 2.0, but it was based only from
one real life workload, so we can expect this to be adjusted in future.
Meanwhile, it's good to have an option to control it and try different
values with different workloads.

The option accepts integer and sets the value of the parameter by
dividing by 10. So the value 20 in the option will set the parameter
value to 2.0.

Any non-negative boost factor is valid (although values above 10.0 are
likely ridiculously large) and should not cause any harm. Value 0 is
certainly valid.

Signed-off-by: Aleksandar Micic <amicic@ca.ibm.com>