Cache resolved methods globally
[skip ci]
Implement global, per RAM class, cross-compilation caching of
resolved methods on the server.
The main challenge with global caching is that resolved methods
are allocated in heap memory, on both the client and the server,
thus they only exist for the duration of one compilation.
We could change allocation to persistent, but that would increase
memory footprint on the client, which should be avoided at all costs.
I addressed this problem by storing data needed to create a resolved
method persistently on the server, and using "on-demand" mirror creation,
i.e. client mirror is recreated only when it's needed.
This global caching is done in addition to local per-compilation caching
that was already being done. I left it in, because local caching is
cheaper, as it doesn't require global lock on RAM class table, and
it doesn't need to recreate mirror on the client.

This is a brief description of the solution I used:
1. Server creates a resolved method for the first time, by making
a remote call to the client.

2. Server saves attributes used to create resolved method
(i.e. `TR_ResolvedJ9JITaaSServerMethodInfo`) in a persistent map
in `ClassInfo`. `ClassInfo` corresponds to class to which this
resolved method belongs. It also saves resolved method into
a local cache.

3. The next time server needs to create the same resolved method,
it will look into the local cache first.
If local cache doesn't have it, server will look into a global cache.
It will find attributes that will be used
to create resolved method without making a remote call.
If a relocatable version of resolved method is required, appropriate
validation record will be created as well.

4. However, _remoteMirror pointer of a newly created resolved method
might be invalid, because client resolved method only exists for the
duration of one compilation. It is set to NULL to indicate that
remote mirror does not exist.

5a. If a method that requires _remoteMirror to exist is called,
we make a remote call to the client and recreate the mirror.

5b. However, there are a few methods that are called very frequently,
so making a remote call every time we need to recreate a mirror
nullifies most of the message savings of global caching.
For such methods, parameters needed for remote mirror creation are
explicitly added to the main message, so recreating a resolved method
will not incur any extra messages.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>