Ensure consistency during relocation failure
No inlined method relocations can be ignore (with or without the SVM)
because of symref sharing. The way the relocation infrastructure ensured
this was by calling the failCompilation method. However, this is not
consistent with the way error handling is performed. Specifically, this
resulted in the compilation thread failing to unreserve the code cache.

This commit ensures that the case in which the relocation infrastructure
cannot ignore a relocation that it cannot perform conforms to the proper
error handling process.

Signed-off-by: Irwin D'Souza <dsouzai@ca.ibm.com>