Implement JIT fields cache
`TR_J9ServerVM::jitFieldsAreSame` compares 2 fields from classes
at given CP indices, and returns true if they are the same,
false otherwise.
We cache the declaring class and value of a field per RAM class,
which enables us to significantly reduce the number of remote calls
made to the client. We only cache fields if they were found, because
if they were not found, they might become visible/resolved in a later
compilation.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>