Make DecimalFormatHelper optional
By checking if the decimal format helper class is available before
running the decimalFormatPeephole optimisation we can make the JIT
tolerate the absence of the class, effectively making the opt switchable
by adding or removing the helper class. This improves flexibility for
deployment and serviceability in the event of needing to revert a
change.

Signed-off-by: James Kingdon <jkingdon@ca.ibm.com>