Reduce IProfiler_persistIprofileInfo messages
[skip ci]
Each inlined method could trigger an `IProfiler_persistIprofileInfo`
update. For each compilation, it would result in more than one
`IProfiler_persistIprofileInfo` messages being sent from the server.
The optimization is to save `IProfiler_persistIprofileInfo` request
during the compilation and send a list of the methods at the end
of the compilation along with `finishCompilation()` message to the client.

Fix #5486

Signed-off-by: Annabelle Huo <Annabelle.Huo@ibm.com>