Ignore current candidate allocation from previous iterations
In TR_EscapeAnalysis::checkAllNewsOnRHSInLoopWithAliasing, while
checking whether all reaching definitions for a particular use of the
current candidate allocation are harmless, code checks whether the
definition referred to one of the candidate allocations, in which case
it was considered harmless.

Previously that was tested by checking whether the RHS of the reaching
definition was the node for a candidate allocation, but that check was
recently changed to also check for its value number.  However, that
allowed a match of the current candidate allocation from a previous
iteration to be considered harmless.

As one of the purposes of the method is to screen out cases in which
the current candidate allocation reaches a subsequent iteration of the
loop, the test of the value number undermines that.  Fixed this by
performing the check of the value numbers of candidate allocations only
for candidates other than the current candidate.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>