Don't check classfiles < 51 for JSRs with -Xverify:none
Classes of version 51 or greater are not allowed to contain JSR/RET
bytecodes.  If running with -Xverify:none, then we don't need to do
the check for them.  Any issues would be reported by enabling the
verifier.

Currently, verification isn't enabled for the bootclasspath so this
change should improve startup for bootclasspath classes

Signed-off-by: Dan Heidinga <daniel_heidinga@ca.ibm.com>