Read new field value before initializing the class
When reporting a static field write from the JIT, the new value pointer
passed in is often a pointer into the java stack. If the report helper
causes the class to be initialized, the stack may grow, invalidating the
pointer.

Read the value from the pointer into a local before initializing the
class.

[ci skip]

Signed-off-by: Graham Chapman <graham_chapman@ca.ibm.com>