Declare J9RAMVirtualMethodRef methodIndexAndArgCount as volatile
Declaring J9RAMVirtualMethodRef.methodIndexAndArgCount as volatile
instructs the compiler to generate assembly code that writes to the
field atomically.
This fixes the segment error due to zero methodIndex which seems due to
separated writes to the field, first one is argCount, and the second one
is methodIndex.
In addition, the field J9RAMVirtualMethodRef.method is declared as
volatile as well.

Signed-off-by: Jason Feng <fengj@ca.ibm.com>