Convert static initialization into a single loop
The code needs to loop over the static fields and initialize any
of them that have a constant value (J9FieldFlagConstant).  Original
code use three loops - one to process objects, one to process single
slot statics, one to process double slot statics - which meant
iterating over the romClass fields 3 times.

The new algorithm uses the counts in the J9ROMClass to figure out
the starting area for each of the 3 kinds of static fields.  This
saves two iterations over the fields (an expensive operation!)
while still maintaining the same layout of object/singles/doubles.

A further optimization to only do this operation if a class has
a static field tagged with J9FieldFlagConstant is possible.

Signed-off-by: Dan Heidinga <daniel_heidinga@ca.ibm.com>