Track constant init static field existence in ClassFileOracle
Keep track of whether the class being processed has static
fields with constantValueAttributes.  These fields require
extra work when initializing the class to ensure that the
constant values are copied into the J9Class->ramStatics.

Having this info allows optimizing class initialization to
skip this processing if the class doesn't have the field

Signed-off-by: Dan Heidinga <daniel_heidinga@ca.ibm.com>