Add libfontconfig1-dev to JDK 8 Dockerfiles
At some point, the JDK 8 extensions repo made a change to its
dependencies and now requires the libfontconfig1-dev package to be
installed. The JDK 8 Dockerfiles in the OpenJ9 repo were never updated
to match this change in dependencies, resulting in Docker builds failing
when running the configuration script. This has now been corrected and
the JDK 8 Dockerfiles should work correctly again.

Signed-off-by: Ben Thomas <ben@benthomas.ca>